Superscalar base file added, minor tweaks
five_stage.c:  adjusted order of command line args to meet the requirements listed

superscalar.c: solves very little of the problem -- currently just blindly jams 2 instructions into whichever pipes and prints out the results. I think this will fail for an odd # of instructions in the trace in its current form. It will take a fair bit of retooling to get it even to meet the second milestone, but I figured I'd make it so we'd have a baseline for building when we meet.