Milestone 3 Draft
I think this should work now. 
The program now accounts for data hazard when packing super instructions and also for load/use hazard for neighboring super instructions. 

TODO: Test files on the unixs cluster to make sure they work in the test environment.
Test cases.
Document & comment source code.
Performance analysis and writeup.

NOTE: The superscalar implementation is not much better than the regular pipeline in this implementation because it so often cannot make use of the duplicate hardware.